% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_utils.R
\name{create_kinfitr_combined_tacs}
\alias{create_kinfitr_combined_tacs}
\title{Create Consolidated Kinfitr Combined TACs Output}
\usage{
create_kinfitr_combined_tacs(
  kinfitr_regions_files_path,
  derivatives_folder,
  output_dir
)
}
\arguments{
\item{kinfitr_regions_files_path}{Path to kinfitr_regions_files.tsv}

\item{derivatives_folder}{Base path to derivatives folder}

\item{output_dir}{Where to save consolidated combined TACs file}
}
\value{
Tibble with all combined TACs data in long format with BIDS attributes
}
\description{
Create single consolidated TSV with BIDS attributes and long-format TACs
}
