% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitrapp-package.R
\docType{package}
\name{kinfitrapp-package}
\alias{kinfitrapp}
\alias{kinfitrapp-package}
\title{kinfitrapp: kinfitr BIDS App Configuration Interface}
\description{
A Shiny web application for creating customized kinfitr BIDS App configuration files for PET imaging analysis. Provides user interface for configuring kinetic modeling parameters for Time Activity Curves (TACs).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
