% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_utils.R
\name{process_all_kinfitr_regions}
\alias{process_all_kinfitr_regions}
\title{Process All kinfitr Regions Across All Files}
\usage{
process_all_kinfitr_regions(
  kinfitr_regions_files_path,
  derivatives_folder,
  output_folder
)
}
\arguments{
\item{kinfitr_regions_files_path}{Path to kinfitr_regions_files.tsv}

\item{derivatives_folder}{Base path to derivatives folder}

\item{output_folder}{Where to save combined TACs files}
}
\value{
Processing summary tibble
}
\description{
Orchestrate entire region combination process across all files
}
