% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_config.R
\name{summarise_tacs_descriptions}
\alias{summarise_tacs_descriptions}
\title{Summarise TACs File Descriptions}
\usage{
summarise_tacs_descriptions(dir_path)
}
\arguments{
\item{dir_path}{Character string path to directory containing *_tacs.tsv files}
}
\value{
Data frame with region configurations from the directory
}
\description{
Process and summarize TACs files descriptions
}
