% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinfitrapp-package.R
\docType{package}
\name{kinfitrapp-package}
\alias{kinfitrapp}
\alias{kinfitrapp-package}
\title{kinfitrapp: Kinfitr BIDS App Configuration Interface}
\description{
A BIDS app implementation and shiny UI for performing PET kinetic modelling using kinfitr. Provides user interface forconfiguring kinetic modeling parameters for Time Activity Curves (TACs).
}
\author{
\strong{Maintainer}: Granville Matheson \email{granville.matheson@ki.se}

Other contributors:
\itemize{
  \item Alessio Giacomel \email{alessio.giacomel@kcl.ac.uk} [contributor]
  \item Martin Noergaard \email{martin.noergaard@di.ku.dk} [contributor]
  \item Anthony Galassi \email{anthony.galassi@nih.gov} [contributor]
}

}
\keyword{internal}
